refactor: move ability system into feature module
