+++ src/features/ability/state.js
@@ -1,0 +1,4 @@
+export const abilityState = {
+  abilityCooldowns: {},
+  actionQueue: [],
+};
