+++ ui/index.js
@@ -50,7 +50,7 @@ import { setupLootUI } from '../src/features/loot/ui/lootTab.js';
 import { renderEquipmentPanel, setupEquipmentTab } from '../src/features/inventory/ui/CharacterPanel.js'; // EQUIP-CHAR-UI
 import { ZONES } from '../data/zones.js'; // MAP-UI-UPDATE
 import { setReduceMotion } from '../src/ui/fx/fx.js';
-import { tickAbilityCooldowns } from '../src/game/abilitySystem.js';
+import { tickAbilityCooldowns } from '../src/features/ability/mutators.js';
 
 // Global variables
 const progressBars = {};
