+++ validation.log
@@ -2,6 +2,37 @@
 
 📋 AI VERIFICATION ENFORCEMENT REPORT
 =====================================
-✅ VERIFICATION PASSED - AI may proceed
+❌ VERIFICATION FAILED - MUST fix before proceeding
+
+🚨 VIOLATIONS DETECTED:
+   • UNDOCUMENTED FILE: src/features/ability/data/abilities.js
+   • UNDOCUMENTED FILE: src/features/ability/logic.js
+   • UNDOCUMENTED FILE: src/features/ability/mutators.js
+   • UNDOCUMENTED FILE: src/features/ability/selectors.js
+   • UNDOCUMENTED FILE: src/features/ability/state.js
+   • UNDOCUMENTED FILE: src/features/affixes/logic.js
+   • UNDOCUMENTED FILE: src/features/affixes/mutators.js
+   • UNDOCUMENTED FILE: src/features/affixes/selectors.js
+   • UNDOCUMENTED FILE: src/features/affixes/state.js
+   • UNDOCUMENTED FILE: src/features/inventory/ui/CharacterPanel.js
+   • UNDOCUMENTED FILE: src/features/inventory/ui/weaponChip.js
+
+📝 REQUIRED ACTION:
+   Update docs/project-structure.md with these files:
+   • src/features/ability/data/abilities.js
+   • src/features/ability/logic.js
+   • src/features/ability/mutators.js
+   • src/features/ability/selectors.js
+   • src/features/ability/state.js
+   • src/features/affixes/logic.js
+   • src/features/affixes/mutators.js
+   • src/features/affixes/selectors.js
+   • src/features/affixes/state.js
+   • src/features/inventory/ui/CharacterPanel.js
+   • src/features/inventory/ui/weaponChip.js
+
+   Then document their purpose and functions.
 
 =====================================
+
+🚫 AI CHANGES BLOCKED UNTIL DOCUMENTATION UPDATED
